slyguy.paramount.plus: Apply request headers when fetching images
The Paramount+ addon was not displaying any sort of images for me.
The Kodi Log contained a lot of errors each time an image was supposed
to be loaded, something like this:

    Error: CCurlFile::Stat - Failed: HTTP response code said error(22)

Interestingly, the logged URLs worked fine on my desktop. I assume that
Paramount+ might have blocked the Kodi user agent at least partially.

After applying this change, the images also started loading in Kodi.