slyguy.paramount.plus: Correctly set locale
Currently most if not all of the descriptions of the Paramount+ Addon
are in English, even if the account is for example German.

The supported locales are part of the client config, with one of them being
marked as default. I believe this is based on geolocation, since the config
is fetched before logging in. After adding it as a parameter to the URL,
all texts in the addon are in the correct language.

I have not tested this extensively, and it might have some edgecases.
It also doesnt allow the user to choose the locale.